create a test that check that a exception was thrown when passing invalid account id

+++ src/main/java/com/example/urbanmobility/service/AccountService.java
@@ -29,6 +29,9 @@ public class AccountService {
     }
 
     public void deleteAccountById(long accountId) {
+        if (!accountRepository.existsById(accountId)){
+            throw new EntityNotFoundException("Account with" + accountId + "does not exist");
+        }
         accountRepository.deleteById(accountId);
     }
 }
